Integrate Sphinx into the build system.
Refs #5774

+++ doc/mkdoc.sh
@@ -1,6 +1,25 @@
 #!/bin/sh
 cd $(dirname -- $0)
 
+if [ -z "$1" ]; then
+  echo "Syntax: $0 <build-dir>"
+  exit 1
+fi
+
+if ! which pandoc; then
+  echo "Please install pandoc to build the documentation files."
+  exit 0
+fi
+
+if ! which pandoc; then
+  echo "Please install sphinx-build to build the documentation files."
+  exit 0
+fi
+
+BUILDDIR="$1"
+
+echo "Build dir: $BUILDDIR"
+
 rm -f index.rst
 cat > index.rst <<RST
 Icinga 2
@@ -26,3 +45,5 @@ for chapter in $(seq 1 100); do
     echo
   done | sed 's/<a id=".*"><\/a>//' | pandoc -f markdown_phpextra -t rst > chapter-$chapter.rst
 done
+
+sphinx-build -b html -d $BUILDDIR/doctrees . $BUILDDIR/htdocs
